Create README.md
People of color are being left behind as the adoption of cryptocurrencies and the baffling blockchain technology captures society spell-bound. Like a speeding train on a lighting railway, many passengers, those not agile enough, get left behind. Minorities are being left behind. Women are being left behind. The techno-challenged are left behind. Many middle-class citizens are left behind. Those with zero disposable income and those hovering near the poverty metric have been left stranded on the platform with their baggage and grip afoot, but without a token of passage, and no ticket agent in sight.

The Blockchain Freedom Foundation was launched in 2021 to establish a pathway for those being left behind to improve their agility in hopping onto the speeding train. We are not suggesting something to impede its trajectory or to slow the train, but rather, this Foundation seeks to establish on-ramps for stranded passengers to safely merge onto the train. 

From observation, the wealth gap between those on board the train and those wandering along the tracks seems to exponentially increase as crypto train bolts from station to station. With every newly coded coin, in most cases, those left behind are the very passengers most needed and most needing to be transported from their current station in life to a new destination of financial freedom. 

Whilst the train is in motion, the Blockchain Freedom Foundation will establish roadmaps for people being left behind to be cautiously on-boarded into the various facets of the crypto concaves. The Foundation will rely on the expertise of early adopters and developers to structure educational tracts about the far-reaching aspects of blockchain technology. In turn, these roadmaps will act as tokens for passage, enabling more and more people to confidently board the train at the nearest station and no longer be left out of the migration.

To enable the onboarding process, this Foundation is engaged in gathering, curating, and breaking down information into consumable nuggets for dissemination. For instance, our websites display a section called “What is” where we use real-life analogies to explain technical and complex blockchain jargon using simple terminology so viewers can get it. 

The object of breaking down the complexity of the information is not to render financial advice or to steer investments toward a particular asset. However, those potentially being left behind can view these bite-sized simplistic messages and infographics on our platforms to quickly acquire basic information of the blockchain technology to make reasonable decisions about what to do with this newly gleaned knowledge they’re absorbing.

In order to achieve the objectives to onboard the targeted groups and the left-behind populous, the Blockchain Freedom Foundation is inviting experts and the wider caring-community from across the color spectrum to join in addressing this major debacle. The community must find a way to onboard those being left out of the cryptocurrency conversation. The efforts of this outreach will result in broader adoption of blockchain as a more transparent system of governance across society and industry.

To advance the mission, the Blockchain Freedom Foundation will employ its media platform to facilitate community actions that bring attention to and attempt to stem systematic exclusion from technological advancements over centuries. The aim is to reverse this persistent tide of leaving people of lesser means out of wealth generation opportunities. 

This platform seeks to provide critical cryptocurrency information and pathways to access economic resources being created by new-age blockchain systems. Equipped with such tools, women, people of color, minority groups, and lower-income citizens can choose to participate or capitalize on opportunities, accumulate store-of-value assets and benefit from potential wealth growth through ownership and investments in cryptocurrencies and blockchain technology.

MISSION - Blockchain Freedom Foundation

To EDUCATE.

To ENCOURAGE debate.

To FOSTER unity.

To BUILD consensus.

To INSPIRE excellence.

FOUNDATION’S CHARTER
The Blockchain Freedom Foundation has embarked on a charter to gather and present knowledge that enables people of all walks of life, but specifically people of color and minority groups, to realize and value their human potential and develop self-sufficiency and independence by accessing and adopting advanced technologies like blockchain and cryptocurrencies.

VISION
The Blockchain Freedom Foundation envisions a community where the adoption of blockchain and other decentralized technologies become the basis of facilitating trust and transparency in all means of communicating, transacting, and ownership worldwide planet earth and intergalactically.

CALL TO ACTION
The Blockchain Freedom Foundation invites corporations, governments, organizations and individuals to recognize the potential that blockchain technology holds for the advancement of people of color, minorities and women; and with that to play an active role in supporting this mission to spread knowledge of advanced technologies like blockchain and cryptocurrencies to people of color so they may participate and enhance their human experience which will result in improving this planetary existence for all.

UTAH BITCOIN CONFERENCE - 10.01.2021

One of the first undertaking by the Blockchain Freedom Foundation is to host the Utah Bitcoin Conference on Friday, October 1st, 2021. We invite the local and international community to come together for this inaugural Bitcoin Conference and Blockchain Party, which will be held at the Zermatt Resort, Midway, Utah, near Park City.

CONFERENCE THEME CLAIM Your Freedom. Secure YOUR Future. Determine Your DESTINY.

This is a gathering of early adopters and Bitcoin enthusiasts from all over the USA and around the globe for this spectacular conference at this spectacular place at this spectacular time, where the Blockchain Freedom Foundation promises a very spectacular experience for all participants.

SPONSORS ARE KEY TO SUCCESS

Your sponsorship of the Utah Bitcoin Conference is key to making the event a success. The Blockchain Freedom Foundation evaluated offering this first conference at no cost to attendees. However, the attractiveness of the venue and opportunity to set the stage for a close-range forum for robust knowledge-filled debate and entertaining discourse was a welcome opportunity. The value sponsors provide allows attendees to receive a deeper and richer experience that only the investment made by your sponsorship could offer at the Utah Bitcoin Conference this year.

EVERYONE CAN SPONSOR

We welcome sponsorship from all who feel inclined to support the Blockchain Freedom Foundation and the Utah Bitcoin Conference. We entertain sponsorships from companies and individuals whose interests align with the vision and any or all objectives of BFF. At this inaugural year of the Utah Bitcoin Conference we are actively soliciting sponsorships and gifts from: Everyday People and Celebrities Corporations who wish to showcase diversity in action Individual Investors Technology Experts Cryptocurrency Exchanges, Token Platforms, Developers Financial institutions, Brokerages, Credit Unions, Banks Utility and Services Sector Builders and Homeowners Universities, Colleges, Educational Entities 

Donations and Sponsorships PAID TO:

Blockchain Freedom Foundation 9740 North 6000 West, Suite 200B, Highland, UT 84003

Contact: 801-299-9105 
EMAIL bitconslc@gmail.com
WEB www.utahbitcoinconference.com